+++ package.json
@@ -18,6 +18,7 @@
     "joi": "^17.9.1",
     "joi-password": "^4.1.1",
     "jsonwebtoken": "^9.0.0",
+    "multer": "^1.4.5-lts.1",
     "nodemailer": "^6.9.1",
     "nodemailer-smtp-transport": "^2.7.4",
     "sequelize": "^6.30.0",
